新增後台member manage

+++ src/main/java/com/finalpretty/app/repositories/MemberRespository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.finalpretty.app.model.Member;
 
-public interface MemberRespository extends JpaRepository<Member, Long> {
+public interface MemberRespository extends JpaRepository<Member, Integer> {
 
 }
